# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Knouz Arabic E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  nameAr: string;
  slug: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  productCount?: number | null;
}

export interface CategoryInput {
  nameAr: string;
  slug: string;
  image?: string;
}

export interface Product {
  id: string;
  nameAr: string;
  descriptionAr: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  images: string[];
  categoryId: string;
  category?: Category;
  featured?: boolean;
  /** @nullable */
  avgRating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt?: string;
}

export interface ProductInput {
  nameAr: string;
  descriptionAr: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  images: string[];
  categoryId: string;
  featured?: boolean;
}

export interface ProductUpdate {
  nameAr?: string;
  descriptionAr?: string;
  price?: number;
  /** @nullable */
  comparePrice?: number | null;
  stock?: number;
  images?: string[];
  categoryId?: string;
  featured?: boolean;
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Review {
  id: string;
  userId?: string;
  /** @nullable */
  userName?: string | null;
  productId: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt?: string;
}

export interface ReviewInput {
  rating: number;
  comment?: string;
  userId?: string;
  userName?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
} as const;

export type OrderShippingAddress = { [key: string]: unknown };

export interface OrderItem {
  id: string;
  productId: string;
  /** @nullable */
  productName?: string | null;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
}

export interface Order {
  id: string;
  /** @nullable */
  orderNumber?: string | null;
  /** @nullable */
  userId?: string | null;
  status: OrderStatus;
  total: number;
  shippingAddress?: OrderShippingAddress;
  paymentMethod?: string;
  paymentStatus?: string;
  /** @nullable */
  couponCode?: string | null;
  discount?: number;
  items: OrderItem[];
  createdAt?: string;
}

export type OrderInputShippingAddress = { [key: string]: unknown };

export interface OrderItemInput {
  productId: string;
  quantity: number;
  price: number;
}

export interface OrderInput {
  userId?: string;
  total: number;
  shippingAddress: OrderInputShippingAddress;
  paymentMethod: string;
  couponCode?: string;
  discount?: number;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  paymentStatus?: string;
}

export interface OrderList {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export type CouponType = typeof CouponType[keyof typeof CouponType];


export const CouponType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED',
} as const;

export interface Coupon {
  id: string;
  code: string;
  discount: number;
  type: CouponType;
  minOrder?: number;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  usageLimit?: number | null;
  usedCount?: number;
}

export type CouponInputType = typeof CouponInputType[keyof typeof CouponInputType];


export const CouponInputType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED',
} as const;

export interface CouponInput {
  code: string;
  discount: number;
  type: CouponInputType;
  minOrder?: number;
  expiresAt?: string;
  usageLimit?: number;
}

export interface CouponValidateInput {
  code: string;
  orderTotal: number;
}

export interface CouponResult {
  valid: boolean;
  discount: number;
  type: string;
  discountAmount: number;
  /** @nullable */
  message?: string | null;
}

export interface UploadInput {
  /** Base64 encoded image data URI */
  data: string;
}

export interface UploadResult {
  url: string;
  /** @nullable */
  publicId?: string | null;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  lowStockCount: number;
  pendingOrders: number;
  recentOrders: Order[];
}

export interface MonthlySales {
  month: string;
  revenue: number;
  orders: number;
}

export type PaymentInitInputBillingData = { [key: string]: unknown };

export interface PaymentInitInput {
  orderId: string;
  amountCents: number;
  billingData: PaymentInitInputBillingData;
}

export interface PaymentInitResult {
  iframeUrl: string;
  /** @nullable */
  paymentKey?: string | null;
}

export type ListProductsParams = {
category?: string;
minPrice?: number;
maxPrice?: number;
sort?: ListProductsSort;
page?: number;
limit?: number;
featured?: boolean;
search?: string;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  popular: 'popular',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
} as const;

export type ListOrdersParams = {
status?: string;
page?: number;
userId?: string;
};

export type TrackOrderParams = {
orderNumber: string;
phone: string;
};

export type GetWishlistParams = {
/**
 * Comma-separated product IDs
 */
ids: string;
};