# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Knouz Arabic E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin = all, user = own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by number and phone
      parameters:
        - name: orderNumber
          in: query
          required: true
          schema:
            type: string
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResult"
        "400":
          description: Invalid or expired coupon
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      responses:
        "200":
          description: Coupons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, status, total, items]
      properties:
        id:
          type: string
        orderNumber:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        status:
          type: string
          enum: [PENDING, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
        total:
          type: number
        shippingAddress:
          type: object
          additionalProperties: true
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        couponCode:
          type: ["string", "null"]
        discount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: ["string", "null"]
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
    OrderInput:
      type: object
      required: [total, items, shippingAddress, paymentMethod]
      properties:
        userId:
          type: string
        total:
          type: number
        shippingAddress:
          type: object
          additionalProperties: true
        paymentMethod:
          type: string
        couponCode:
          type: string
        discount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity, price]
      properties:
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: number
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
        paymentStatus:
          type: string
    OrderList:
      type: object
      required: [orders, total, page, totalPages]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    Coupon:
      type: object
      required: [id, code, discount, type]
      properties:
        id:
          type: string
        code:
          type: string
        discount:
          type: number
        type:
          type: string
          enum: [PERCENTAGE, FIXED]
        minOrder:
          type: number
        expiresAt:
          type: ["string", "null"]
        usageLimit:
          type: ["integer", "null"]
        usedCount:
          type: integer
    CouponInput:
      type: object
      required: [code, discount, type]
      properties:
        code:
          type: string
        discount:
          type: number
        type:
          type: string
          enum: [PERCENTAGE, FIXED]
        minOrder:
          type: number
        expiresAt:
          type: string
        usageLimit:
          type: integer
    CouponValidateInput:
      type: object
      required: [code, orderTotal]
      properties:
        code:
          type: string
        orderTotal:
          type: number
    CouponResult:
      type: object
      required: [valid, discount, type, discountAmount]
      properties:
        valid:
          type: boolean
        discount:
          type: number
        type:
          type: string
        discountAmount:
          type: number
        message:
          type: ["string", "null"]
...