# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Knouz Arabic E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderShippingAddress } from './orderShippingAddress';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  /** @nullable */
  orderNumber?: string | null;
  /** @nullable */
  userId?: string | null;
  status: OrderStatus;
  total: number;
  shippingAddress?: OrderShippingAddress;
  paymentMethod?: string;
  paymentStatus?: string;
  /** @nullable */
  couponCode?: string | null;
  discount?: number;
  items: OrderItem[];
  createdAt?: string;
}